# Enable fully threaded IO for installs