--- a/README.md
+++ b/README.md
@@ -710,6 +710,9 @@ yet validate signatures of downloads.
 
 [s]: https://github.com/rust-lang/rustup.rs/issues?q=is%3Aopen+is%3Aissue+label%3Asecurity
 
+File modes on installation honor umask as of 1.18.4, use umask if
+very tight controls are desired.
+
 ## FAQ
 
 ### Is this an official Rust project?