--- a/src/diskio/immediate.rs
+++ b/src/diskio/immediate.rs
@@ -4,8 +4,6 @@
 /// threaded code paths.
 use super::{perform, Executor, Item};
 
-use std::cell::Cell;
-
 pub struct ImmediateUnpacker {}
 impl ImmediateUnpacker {
     pub fn new<'a>() -> ImmediateUnpacker {
@@ -13,40 +11,17 @@ impl ImmediateUnpacker {
     }
 }
 
-enum IterateOne {
-    Item(Item),
-    None,
-}
-
-impl Default for IterateOne {
-    fn default() -> Self {
-        IterateOne::None
-    }
-}
-
-struct ImmediateIterator(Cell<IterateOne>);
-
-impl Iterator for ImmediateIterator {
-    type Item = Item;
-    fn next(&mut self) -> Option<Item> {
-        match self.0.take() {
-            IterateOne::Item(item) => Some(item),
-            IterateOne::None => None,
-        }
-    }
-}
-
 impl Executor for ImmediateUnpacker {
     fn dispatch(&mut self, mut item: Item) -> Box<dyn '_ + Iterator<Item = Item>> {
         perform(&mut item);
-        Box::new(ImmediateIterator(Cell::new(IterateOne::Item(item))))
+        Box::new(Some(item).into_iter())
     }
 
-    fn join(&mut self) -> Option<Box<dyn Iterator<Item = Item>>> {
-        None
+    fn join(&mut self) -> Box<dyn Iterator<Item = Item>> {
+        Box::new(None.into_iter())
     }
 
-    fn completed(&mut self) -> Option<Box<dyn Iterator<Item = Item>>> {
-        None
+    fn completed(&mut self) -> Box<dyn Iterator<Item = Item>> {
+        Box::new(None.into_iter())
     }
 }