… Cargo.lock
… src/dist/component/package.rs
… tests/dist_install.rs