See the individual commits for what this patch does, but in short on all platforms it will use  up to CPU-count threads during unpacking, will no longer do a separate pass to set permissions, and may use more memory. Uninstall hasn't been optimised yet (and upgrades perform uninstalls, so the instructions below separate the steps for better reporting of results)

I'd like to get some wider testing that we've had so far. I can supply test binaries if needed, but its very easy to test with:
```
git clone https://github.com/rust-lang/rustup.rs.git
cd rustup.rs
git fetch origin pull/1876/head:threadedio
git checkout threadedio
cargo build --release
cp target/release/rustup{-init,} #.exe if you are on windows
./target/release/rustup uninstall nightly
./target/release/rustup/install nightly
```

Let us know how this goes. I'm particularly interested in knowing any *regressions* that would make it unsuitable for merging.

This is known not to solve WSL performance issues: #1866.
It is thought to help with NFS issues #1867 and should further help windows installation times #1540.