@kinnison 2019-05-26
What does this mean in terms of max open handles to close?  is that limited to the number of hardware threads too?