@rbtcollins 2019-05-26
Nope. I will get to that. I'd like some HPC folk to test this as is though

On Mon, 27 May 2019, 03:06 Daniel Silverstone, <notifications@github.com>
wrote:

> What does this mean in terms of max open handles to close? is that limited
> to the number of hardware threads too?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rustup.rs/pull/1876?email_source=notifications&email_token=AADZ7XXVFQYW6IT7YFEGHV3PXKROTA5CNFSM4HPWJBRKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODWIHTOQ#issuecomment-496007610>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADZ7XTDGIF4U6HSIPKA7KTPXKROTANCNFSM4HPWJBRA>
> .
>