@kinnison 2019-05-26
I assume you'll be recommending they build a version to trial from your branch then?  If you want it merged before then, then we should open an issue to block 1.19.0 to ensure we don't forget to get the rlimits or similar in place should we not back this out.