@kinnison 2019-05-26
Also, on the off-chance that we can work out why appveyor is squiffing, if you rebase onto master, we have `RUST_BACKTRACE=1` set for at least some of the testing.  It may not get everything but it might help a bit.