@saethlin 2019-05-26
I'd like to test this but I'm not really sure _how_. I've cloned @rbtcollins bug-1866, but when I build the only binary I get is `rustup-init`. Do I have to uninstall my other rustup then re-init the whole thing with this?