@kinnison 2019-05-26
@saethlin The binary is called `rustup-init` you can install it without having to uninstall your current version by running it directly.  For testing locally I simply to `cargo run -- -v -y --no-modify-path`.

Once you've done that it should have installed into the usual place (`~/.cargo.bin`) and you can run `rustup toolchain install --no-self-update nightly` or somesuch to test the speed.  Remember the `--no-self-update` or else it'll download and replace itself with the release version after installing the channel you select.

Once you're done with your testing you can simply do `rustup self update` and it'll download and install the current release instead of your locally built copy.