@rbtcollins 2019-05-26
What I normally do is build it with cargo build ---release, then cp `target/release/rustup{-init,}` and finally ./target/release/rustup install nightly (or whatever) - that way I don't care what rustup is installed.