@rbtcollins 2019-05-26
Thanks for testing!

On Mon, 27 May 2019, 09:47 Ben Kimock, <notifications@github.com> wrote:

> No improvement, ~5 second regression but that's withing measurement
> uncertainty here.
>
> info: syncing channel updates for 'nightly-x86_64-unknown-linux-gnu'
> info: latest update on 2019-05-26, rust version 1.36.0-nightly (f49269398 2019-05-25)
> info: downloading component 'rustc'
>  86.4 MiB /  86.4 MiB (100 %)  24.4 MiB/s in  3s ETA:  0s
> info: downloading component 'rust-std'
>  60.7 MiB /  60.7 MiB (100 %)  23.5 MiB/s in  2s ETA:  0s
> info: downloading component 'cargo'
> info: downloading component 'rust-docs'
> info: installing component 'rustc'
>  86.4 MiB /  86.4 MiB (100 %)   8.8 MiB/s in  9s ETA:  0s
> info: installing component 'rust-std'
>  60.7 MiB /  60.7 MiB (100 %)   9.1 MiB/s in  7s ETA:  0s
> info: installing component 'cargo'
> info: installing component 'rust-docs'
>  11.0 MiB /  11.0 MiB (100 %) 443.2 KiB/s in  2m  3s ETA:  0s
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rustup.rs/pull/1876?email_source=notifications&email_token=AADZ7XQDAT72RCRMHQ37UG3PXMAPTA5CNFSM4HPWJBRKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODWIOB7Y#issuecomment-496034047>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADZ7XRAUYIIZLGVIXB63VLPXMAPTANCNFSM4HPWJBRA>
> .
>