@saethlin 2019-05-30
Looks like a substantial improvement, and I can't detect any trampling. Nice!
```
info: syncing channel updates for 'nightly-x86_64-unknown-linux-gnu'
info: latest update on 2019-05-30, rust version 1.37.0-nightly (37d001e4d 2019-05-29)
info: downloading component 'rustc'
 86.0 MiB /  86.0 MiB (100 %)  60.6 MiB/s in  1s ETA:  0s
info: downloading component 'rust-std'
 60.4 MiB /  60.4 MiB (100 %)  28.8 MiB/s in  2s ETA:  0s
info: downloading component 'cargo'
info: downloading component 'rust-docs'
info: installing component 'rustc'
 86.0 MiB /  86.0 MiB (100 %)  13.9 MiB/s in  6s ETA:  0s
info: installing component 'rust-std'
 60.4 MiB /  60.4 MiB (100 %)  16.9 MiB/s in  3s ETA:  0s
info: installing component 'cargo'
info: installing component 'rust-docs'
 11.0 MiB /  11.0 MiB (100 %) 233.6 KiB/s in 15s ETA:  0s
```