@rbtcollins 2019-05-30
I'm going to add more instrumentation and so forth before we merge this, but you're welcome to use this branch for now :)