@Coder-256 2019-06-01
I wonder if "WslFs" might have something to do with this? [More info](https://devblogs.microsoft.com/commandline/whats-new-for-wsl-in-windows-10-version-1903/#comments-2438)