@rbtcollins 2019-06-06
> I wonder if "WslFs" might have something to do with this? [More info](https://devblogs.microsoft.com/commandline/whats-new-for-wsl-in-windows-10-version-1903/#comments-2438)

I don't think so: the plan9 fs stuff is super interesting but thats for interop with windows mounted dirs as opposed to working with the WSL owned dirs, isn't it?