@rbtcollins 2019-06-07
@Coder-256 I wish it was :) - Just coincidentally the MS Defender team have released signature updates about now hinting to defender how to deal with rustup  under WSL similarly to rustup on Windows (scan on read rather than scan on close) - this has massively reduced the duration of the CloseHandle syscall and that improves performance the same as it did creating a manual Defender exclusion.

Current rustup in my *not upgraded* WSLv1: 
```info: installing component 'rust-docs'
 11.2 MiB /  11.2 MiB (100 %) 329.6 KiB/s in 27s ETA:  0s
```
This patch:
```info: installing component 'rust-docs'
 11.2 MiB /  11.2 MiB (100 %) 132.8 KiB/s in 20s ETA:  0s
```
Three days ago, nothing else changed except defender signature updates...
```info: installing component 'rust-docs'
 11.2 MiB /  11.2 MiB (100 %) 419.2 KiB/s in  2m 50s ETA:  0s
```